+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/illegalthrows/InputIllegalThrows.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.coding;
+package com.puppycrawl.tools.checkstyle.checks.coding.illegalthrows;
 /** Input file */
 public class InputIllegalThrows {
