+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/illegalthrows/InputIllegalThrowsIgnoreOverriddenMethods.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.coding;
+package com.puppycrawl.tools.checkstyle.checks.coding.illegalthrows;
 
 public class InputIllegalThrowsIgnoreOverriddenMethods
              extends InputIllegalThrows
